Add viewing and creating tweets. Fix layout issue.
This commit adds various progress in different areas.

There is a layout issue inherent in NextJS and Material-UI based
projects. The styles for MUI are not loaded correctly with SSR. After
enough searching online, I was able to find an example Next project by
the MUI folks. The changes in `_document.js` solve the layout issue via
Next's `enhanceApp` parameter prop on `ctx.renderPage`.

Writing tweets is done through the <CreateTweet /> component which will
allow users to append images and possibly other media in the future.
I have separated that component from the overall Timeline view as it is
its own component.

The `Tweet` interface has been added. I'm unsure exactly which
properties will stay on that interface and which will go. `verified` is
a bit unnecessary but may be desirable for users.

This commit also adds the `maxWidth` variable to `Columns` and `Rows`.