[WIP] Can preview image upload before tweeting
This commit starts work on supporting images in tweets. Tweets will not
post with the attached image just yet. They are only previewed thus far.

It appears that `URL.createObjectURL` is not supported. If it did exist,
it would be easier to convert `mediaFile` to a string. Then that string
is fed into something as the image preview. `FileReader` is the
alternative. It may actually be better due to FileReader exposing
loading state I can hook UI into.

This commit also fixed an issue with firebase initializing the `app`
twice. Don't know why that was a problem in the first place or why it
was not encountered in previous Firebase projects.

This commit also fixes a silly z-index issue with the sticky "Home"
header above the Timeline.