WIP Can tweet with image upload
This commit is a work in progress effort on making tweets
postable with an image attachment. The issue encountered is Firebase
requires the client that is uploading to storage to be authenticated
first.

The user authentication story/feature must be completed before image
upload can work.

This commit also makes the "Tweet" button in `<CreateTweet />` work.

This commit also orders the tweets in the timeline by the latest tweets
up at the top.